requirements.txt
PyQt6>=6.4.0
python-chess>=1.9.0
lmdb>=1.3.0
requests>=2.28.0
aiohttp>=3.8.0
asyncio-mqtt>=0.11.0
numpy>=1.21.0
pandas>=1.5.0
flask>=2.3.0
flask-cors>=4.0.0
zstandard>=0.21.0 